+++ src/services/authenticationService/authentications/index.ts
@@ -1,2 +1,3 @@
 export * from './createBasicAuthenticationToken';
+export * from './createJWTAuthentication';
 export * from './createOAuth2AuthenticationToken';